11 create pg function calling template (#19)
* implement user profile update functionality

* Add user watchlist functionality and related database changes

* Refactor watchlist models and update related tests; add SQL query for string search

* Add user watchlist retrieval endpoint and related model; enhance error handling

* Add average rating field to movie, episode, and series models; update database schema and related functions

* Refactor database functions and triggers for user ratings; update drop scripts and add average rating calculations

* Add user rating models and related database functions; update context and drop scripts

* Add ratings episode schema and update related tests; refactor watchlist request class

* Add ID validation service and related schemas; refactor user controller and tests

* Enhance ID validation and update RatingsController with detailed API endpoint comments

* Fix nullability issue in RatingsApiTest by adding null-forgiving operator to deserialization

* Remove obsolete functions related to movie average rating updates from SQL scripts

* Add get_user_recent_view function and related tests; remove obsolete recent view function

* Add RecentView and RecentViewAll models; update OMGdbContext and clean up using directives

* Refactor using directives in Watchlist and RecentView schemas; add RecentViewSchema class for testing

+++ OMGdbApi/Controllers/UserController.cs
@@ -20,7 +20,7 @@ namespace OMGdbApi.Controllers
     public class UserController : ControllerBase
     {
         private readonly OMGdbContext _context;
-        private readonly Hashing _hasing = new Hashing();
+        private readonly Hashing _hasing = new();
 
         public UserController(OMGdbContext context, Hashing hasing)
         {
@@ -64,12 +64,25 @@ namespace OMGdbApi.Controllers
         [HttpGet("{id}")]
         [Authorize]
         public async Task<ActionResult<UserDTO>> GetUser(string id)
-        {
+        {   
+            if (string.IsNullOrEmpty(id))
+            {
+                return BadRequest();
+            }
+
+        
             var user = await _context.Users.FindAsync(id);
 
             if (user == null)
             {
                 return NotFound();
+            }
+
+            var token_id = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            
+            if (token_id != user.Id)
+            {
+                return Unauthorized();
             }
 
             return UserDTO(user);
@@ -80,10 +93,12 @@ namespace OMGdbApi.Controllers
         [HttpPut("{id}")]
         [Authorize]
         public async Task<ActionResult<UserDTO>> PutUser(string id, string name, string email)
-        {
-            if (string.IsNullOrEmpty(name) || string.IsNullOrEmpty(email))
+        {   
+            
+
+            if (string.IsNullOrEmpty(name) || string.IsNullOrEmpty(email) || string.IsNullOrEmpty(id))
             {
-                return BadRequest();
+                return BadRequest("Name, email or id is null");
             }
 
             
@@ -91,19 +106,26 @@ namespace OMGdbApi.Controllers
 
             if (user == null)
             {
-                return NotFound();
-            }    
+                return NotFound("User not found");
+            } 
+
+            var token_id = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            
+            if (token_id != user.Id)
+            {
+                return Unauthorized("Unauthorized");
+            }
+
             user.Name = name;
             user.Email = email;
             
-
             try
             {
                 await _context.SaveChangesAsync();
             }
             catch (DbUpdateConcurrencyException) when (!UserExists(id))
             {
-               return NotFound();
+               return NotFound("User not found");
             }
 
             return CreatedAtAction(nameof(GetUser), new { id = user.Id }, UserDTO(user));
@@ -122,7 +144,7 @@ namespace OMGdbApi.Controllers
                 || string.IsNullOrEmpty(userCreate.Email)
             )
             {
-                return BadRequest();
+                return BadRequest("Name, password or email is null");
             }
 
             // Check if email already exists
@@ -152,7 +174,7 @@ namespace OMGdbApi.Controllers
             {
                 if (user.Id != null && UserExists(user.Id))
                 {
-                    return Conflict();
+                    return Conflict("User already exists.");
                 }
                 else
                 {
@@ -225,7 +247,7 @@ namespace OMGdbApi.Controllers
                 issuer: "OMGdbApi",
                 audience: "OMGdbApi",
                 claims: claims,
-                expires: DateTime.Now.AddMinutes(30),
+                expires: DateTime.Now.AddMinutes(1440),
                 signingCredentials: creds
             );
 
@@ -243,11 +265,23 @@ namespace OMGdbApi.Controllers
         [HttpDelete("{id}")]
         [Authorize]
         public async Task<IActionResult> DeleteUser(string id)
-        {
+        {   
+            if (string.IsNullOrEmpty(id))
+            {
+                return BadRequest();
+            }
+
             var user = await _context.Users.FindAsync(id);
             if (user == null)
             {
                 return NotFound();
+            }
+
+            var token_id = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            
+            if (token_id != user.Id)
+            {
+                return Unauthorized();
             }
 
             _context.Users.Remove(user);
@@ -268,7 +302,7 @@ namespace OMGdbApi.Controllers
                 Name = user.Name,
                 Email = user.Email,
                 Created_at = user.Created_at,
-            };
+            };   
 
     }
 }
